rename abspkg -> abcpkg
